chore: moved actions into own file, state-context renamed to index
