+++ src/context/actions/infoStrip.js
