+++ src/context/actions/updateHtml.js
@@ -1,14 +1,14 @@
-import InfoStripTemplate from '../components/InfoStrip/InfoStripTemplate'
-import HeroTemplate from '../components/Hero/HeroTemplate'
-import LowerTemplate from '../components/Lower/LowerTemplate'
-import SaleTemplate from '../components/Sale/SaleTemplate'
-import SaleCategoriesTemplate from '../components/SaleCategories/SaleCategoryTemplate'
-import MegaBannerWidget from '../components/MegaBanner'
-import TickerTemplate from '../components/Ticker/TickerTemplate'
-import CategoryTemplate from '../components/Categories/CategoryTemplate'
-import addPlaceholderImageCSS from '../lib/addPlaceholderImageCSS'
-import { AppDownloadTemplate } from '../components/AppDownload'
-import { SaleTextBannerTemplate } from '../components/SaleTextBanner'
+import InfoStripTemplate from '../../components/InfoStrip/InfoStripTemplate'
+import HeroTemplate from '../../components/Hero/HeroTemplate'
+import LowerTemplate from '../../components/Lower/LowerTemplate'
+import SaleTemplate from '../../components/Sale/SaleTemplate'
+import SaleCategoriesTemplate from '../../components/SaleCategories/SaleCategoryTemplate'
+import MegaBannerWidget from '../../components/MegaBanner'
+import TickerTemplate from '../../components/Ticker/TickerTemplate'
+import CategoryTemplate from '../../components/Categories/CategoryTemplate'
+import addPlaceholderImageCSS from '../../lib/addPlaceholderImageCSS'
+import { AppDownloadTemplate } from '../../components/AppDownload'
+import { SaleTextBannerTemplate } from '../../components/SaleTextBanner'
 
 const hasContent = (state) => {
   if (Array.isArray(state)) {
